% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosePrice.R
\name{getClosePrice}
\alias{getClosePrice}
\title{getClosePrice Barchart data.}
\usage{
getClosePrice(symbols, date = NULL, splits = 1, dividends = 1)
}
\arguments{
\item{symbols}{required; A symbol or code that identifies a financial instrument. Multiple symbols separated by a comma may be used. Example IBM,AAPL,GOOG}

\item{date}{optional; The date of the historical data query. The value should conform to the format yyyymmdd. Example 2010-01-01}

\item{splits}{optional; An adjustment of stock value due to corporate action. This parameter only applies to stocks and specifies whether the data returned should be adjusted for splits or not. Set to true to query for adjusted the data, or to false for non-adjusted data. Default 1. Example true}

\item{dividends}{optional; A distribution of a portion of a company's earnings. This parameter only applies to stocks and specifies whether the data returned should be adjusted for dividends or not. Set to true to query for adjusted the data, or to false for non-adjusted data. Default 1. Example true}
}
\value{
data frame
}
\description{
\code{getClosePrice} returns an data frame
The getClosePrice API provides the close price for given instruments for the given date.
}
\examples{
\donttest{
getClosePrice(symbols = 'ESM19')
getClosePrice(symbols = 'NGM19,ESM19')
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getFuturesExpirations}
}
\author{
Mario Pisa
}
