% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuturesExpirations.R
\name{getFuturesExpirations}
\alias{getFuturesExpirations}
\title{getFuturesExpirations Barchart data.}
\usage{
getFuturesExpirations(roots = NULL, contracts = NULL)
}
\arguments{
\item{roots}{optional; A symbol or code that identifies a futures root. Example ZC,ZS}

\item{contracts}{optional; A symbol or code that identifies a futures contract.}
}
\value{
data frame
}
\description{
\code{getFuturesExpirations} returns an data frame
The getFuturesExpiration API from Barchart OnDemand provides first notice and last trade dates for futures contracts.
}
\examples{
\donttest{
getFuturesExpirations(roots = 'ES')
getFuturesExpirations(contracts = 'NGM19,ESM19')
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getFuturesExpirations}
}
\author{
Mario Pisa
}
