% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuturesSpecifications.R
\name{getFuturesSpecifications}
\alias{getFuturesSpecifications}
\title{getFuturesSpecifications Barchart data.}
\usage{
getFuturesSpecifications(symbols, exchanges = NULL, futuresCategory = NULL)
}
\arguments{
\item{symbols}{optional; A symbol or code that identifies a financial instrument. Multiple symbols separated by a comma may be used. Example GC,CL}

\item{exchanges}{optional; The code for the exchange a group of financial instruments are listed on. Example COMEX,NYMEX}

\item{futuresCategory}{optional; Futures categories include: Energies, Financials, Grains, Indices, Meats, Metals, Softs and Currencies. Example energies,metals}
}
\value{
data frame
}
\description{
\code{getFuturesSpecifications} returns an data frame
Barchart OnDemand's getFuturesSpecifications API provides contract information such as trading hours, contract size and tick size for futures contracts.
}
\examples{
\donttest{
getFuturesSpecifications(symbols = 'CL')
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getFuturesSpecifications}
}
\author{
Mario Pisa
}
