% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuote.R
\name{getQuote}
\alias{getQuote}
\title{getQuote Barchart data.
The getQuote API is used to request price data, either real-time,
   delayed or end-of-day, by symbol on stocks, indexes, mutual funds, ETFs,
   futures, foreign exchange, or cryptocurrencies.}
\usage{
getQuote(symbols, fields = NULL, only = NULL)
}
\arguments{
\item{symbols}{Required; A symbol or code that identifies a financial instrument. Multiple symbols separated by a comma may be used. For futures, notation such as for the active contract or for the first nearby is supported, as is for all futures contracts for a given root symbol, as is for all options contracts for a given underlying futures symbol.}

\item{fields}{optional; The fields requested. Example fiftyTwoWkHigh,fiftyTwoWkHighDate,fiftyTwoWkLow,fiftyTwoWkLowDate}

\item{only}{optional; Returns only specified fields. Example symbol,name}
}
\value{
xts
}
\description{
\code{getQuote} returns an xts.
The data has:
- Tot
}
\examples{
getQuote(symbols = 'ESM19')
getQuote(symbols = 'NGM19,ESM19', fields = 'openInterest')
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getQuote}
}
\author{
Mario Pisa
}
