% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuoteEod.R
\name{getQuoteEod}
\alias{getQuoteEod}
\title{getQuoteEod Barchart data.}
\usage{
getQuoteEod(symbols, exchange = NULL)
}
\arguments{
\item{symbols}{required; An MIC code or a Country code plus the symbol or code that identifies a financial instrument. Multiple symbols separated by a comma may be used. For futures, notation such as for the active contract or for the first nearby is supported, as is for all futures contracts for a given root symbol. If no country code or MIC is passed then the API will behave the same as the getQuote API. Example XASE:SPY,US:IBM}

\item{exchange}{optional; The list of valid exchange codes to limit symbol search. Example NYSE,AMEX}
}
\value{
data frame
}
\description{
\code{getQuoteEod} returns an data frame
The getQuoteEod API is used to request end-of-day price data,
by combined exchange and symbol, on stocks, indexes, mutual funds, ETFs,
futures, foreign exchange, or cryptocurrencies. Historical data is
available as tick, minute bars or end-of-day data.
}
\examples{
\donttest{
getQuoteEod(symbols = 'XASE:SPY')
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getQuoteEod}
}
\author{
Mario Pisa
}
