% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbolLookup.R
\name{getSymbolLookUp}
\alias{getSymbolLookUp}
\title{getSymbolLookup Barchart data.}
\usage{
getSymbolLookUp(keyword, limit = 10, exchanges = "NYSE,AMEX,NASDAQ,COMEX",
  moreBestMatches = "true", moreBestMatchLimit = 4, fields = "")
}
\arguments{
\item{keyword}{Required; The keyword on which the search needs to be made.}

\item{limit}{optional; The number of results to be returned.}

\item{exchanges}{optional; Valid exchange code. Multiple exchanges can be passed separated by a comma. Default NYSE,AMEX,NASDAQ,COMEX}

\item{moreBestMatches}{optional; To get additional best match results. Boolean. Dafault = 'true'.}

\item{moreBestMatchLimit}{optional; The number of additional best match results that should be returned. Default = 4}

\item{fields}{optional; The fields requested. A comma or semi-colon delimited string.}
}
\value{
data frame
}
\description{
\code{getSymbolLookUp} returns an data frame.
The getSymbolLookUp API provides a list of all instruments based on the keyword being passed.
}
\examples{
\donttest{
getSymbolLookUp('GC', limit = 100, exchanges = 'COMEX')
getSymbolLookUp('Gold', limit = 100, exchanges = 'COMEX')
getSymbolLookUp('Gold', limit = 100, exchanges = 'CXMI')
getSymbolLookUp('Zinc', limit = 1000, exchanges = 'COMEX')
getSymbolLookUp('A', limit = 1000, exchanges = 'COMEX')
getSymbolLookUp('E-Mini', limit = 100, exchanges = 'CXMI')
getSymbolLookUp('E-Micro', limit = 100, exchanges = 'CXMI')
getSymbolLookUp('19', limit = 100, exchanges = 'CXMI')
getSymbolLookUp('19', limit = 100, exchanges = 'COMEX')
getSymbolLookUp('\%', limit = 90000, exchanges = 'COMEX', moreBestMatches = 'true')
getSymbolLookUp('\%', limit = 90000, exchanges = 'CXMI')
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api/getSymbolLookUp}
}
\author{
Mario Pisa
}
