% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPIkey.R
\name{setAPIkey}
\alias{setAPIkey}
\title{setAPIkey Barchart.
The setAPIkey is used to setup the Barchart API key.}
\usage{
setAPIkey(apikey, premium = FALSE)
}
\arguments{
\item{apikey}{required; Barchart API key}

\item{premium}{optional; default = FALSE}
}
\value{
xts
}
\description{
\code{setAPIkey} returns an xts.
The data has:
- Tot
}
\examples{
setAPIkey(key = '<Your Barchart API key>')
setAPIkey(key = '<Your Barchart API key>', premium = TRUE)
}
\seealso{
\url{https://www.barchart.com/ondemand/api}
}
\author{
Mario Pisa
}
