% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPIkey.R
\name{setAPIkey}
\alias{setAPIkey}
\title{setAPIkey Barchart.}
\usage{
setAPIkey(apikey, premium = FALSE)
}
\arguments{
\item{apikey}{required; Barchart API key}

\item{premium}{optional; default = FALSE}
}
\value{
var
}
\description{
\code{setAPIkey} returns a hide APIkey variable.
The setAPIkey is used to setup the Barchart API key.
}
\examples{
\donttest{
setAPIkey(apikey = '<Your Barchart API key>')
setAPIkey(apikey = '<Your Barchart API key>', premium = TRUE)
}
}
\seealso{
\url{https://www.barchart.com/ondemand/api}
}
\author{
Mario Pisa
}
